requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
# altgraph=0.17.3=pypi_0
appdirs=1.4.4=pyhd3eb1b0_0
blas=1.0=mkl
bottleneck=1.3.5=py39h080aedc_0
# brewer2mpl=1.4.1=pypi_0
brotli=1.0.9=h8ffe710_7
brotli-bin=1.0.9=h8ffe710_7
brotlipy=0.7.0=py39h2bbff1b_1003
ca-certificates=2023.05.30=haa95532_0
certifi=2023.7.22=py39haa95532_0
cffi=1.15.1=py39h2bbff1b_3
charset-normalizer=2.0.4=pyhd3eb1b0_0
# click=8.1.3=pypi_0
# colorama=0.4.6=pypi_0
cryptography=39.0.1=py39h21b164f_0
cycler=0.11.0=pyhd8ed1ab_0
eigen=3.4.0=h2d74725_0
# et-xmlfile=1.1.0=pypi_0
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.10.4=h546665d_1
glib=2.69.1=h5dc1a3c_2
gst-plugins-base=1.18.5=h9e645db_0
gstreamer=1.18.5=hd78058f_0
hdf5=1.10.6=nompi_he0bbb20_101
icc_rt=2022.1.0=h6049295_2
icu=58.2=ha925a31_3
idna=3.4=py39haa95532_0
intel-openmp=2021.4.0=haa95532_3556
jpeg=9e=h8ffe710_2
kiwisolver=1.4.4=py39h2e07f2f_0
libbrotlicommon=1.0.9=h8ffe710_7
libbrotlidec=1.0.9=h8ffe710_7
libbrotlienc=1.0.9=h8ffe710_7
libclang=12.0.0=default_h627e005_2
libffi=3.4.2=h8ffe710_5
libiconv=1.17=h8ffe710_0
libogg=1.3.4=h8ffe710_1
libpng=1.6.37=h2a8f88b_0
libprotobuf=3.20.3=h23ce68f_0
libtiff=4.0.9=h36446d0_1002
libvorbis=1.3.7=h0e60522_0
libwebp=1.2.4=h8ffe710_0
libwebp-base=1.2.4=h8ffe710_0
libxml2=2.9.14=h0ad7f3c_0
libxslt=1.1.35=h2bbff1b_0
lz4-c=1.9.3=h8ffe710_1
matplotlib=3.5.3=py39hcbf5309_2
matplotlib-base=3.5.3=py39h581301d_0
mkl=2021.4.0=haa95532_640
mkl-service=2.4.0=py39h2bbff1b_0
mkl_fft=1.3.1=py39h277e83a_0
mkl_random=1.2.2=py39hf11a4ad_0
munkres=1.1.4=pyh9f0ad1d_0
numexpr=2.8.4=py39h5b0cc5e_0
numpy=1.23.5=py39h3b20f71_0
numpy-base=1.23.5=py39h4da318b_0
opencv=4.6.0=py39h104de81_2
# openpyxl=3.1.2=pypi_0
openssl=1.1.1v=h2bbff1b_0
packaging=23.0=pyhd8ed1ab_0
pandas=1.5.3=py39hf11a4ad_0
pcre=8.45=h0e60522_0
# pefile=2023.2.7=pypi_0
pillow=9.3.0=py39hdc2b20a_1
pip=22.3.1=py39haa95532_0
ply=3.11=py_1
pooch=1.4.0=pyhd3eb1b0_0
pycparser=2.21=pyhd3eb1b0_0
# pyinstaller=5.12.0=pypi_0
# pyinstaller-hooks-contrib=2023.3=pypi_0
pyopenssl=23.0.0=py39haa95532_0
pyparsing=3.0.9=pyhd8ed1ab_0
# pyqt5=5.15.9=pypi_0
# pyqt5-plugins=5.15.9.2.3=pypi_0
# pyqt5-qt5=5.15.2=pypi_0
pyqt5-sip=12.11.0=py39hd77b12b_0
# pyqt5-tools=5.15.9.3.3=pypi_0
pysocks=1.7.1=py39haa95532_0
python=3.9.16=h6244533_0
python-dateutil=2.8.2=pyhd8ed1ab_0
# python-dotenv=1.0.0=pypi_0
python_abi=3.9=2_cp39
pytz=2022.7=py39haa95532_0
# pywin32-ctypes=0.2.1=pypi_0
# qt-main=5.15.2=he8e5bd7_7
# qt-webengine=5.15.9=hb9a9bb5_5
# qt5-applications=5.15.2.2.3=pypi_0
# qt5-tools=5.15.2.1.3=pypi_0
# qtwebkit=5.212=h3ad3cdb_4
requests=2.28.1=py39haa95532_1
scipy=1.10.1=py39h321e85e_0
# seaborn=0.12.2=pypi_0
setuptools=65.6.3=py39haa95532_0
sip=6.6.2=py39h415ef7b_0
six=1.16.0=pyh6c4a22f_0
sqlite=3.40.1=h2bbff1b_0
tk=8.6.12=h8ffe710_0
toml=0.10.2=pyhd8ed1ab_0
tornado=6.2=py39hb82d6ee_0
tzdata=2022g=h04d1e81_0
urllib3=1.26.15=py39haa95532_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
wheel=0.37.1=pyhd3eb1b0_0
win_inet_pton=1.1.0=py39haa95532_0
wincertstore=0.2=py39haa95532_2
zlib=1.2.13=h8cc25b3_0
zstd=1.5.2=h19a0ad4_0
